pyproject.toml
[project]
name = "genoshi"
version = "0.1.0"
description = "Mini Insurance Document Validator - AI-powered document validation API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.1",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
    "uvicorn[standard]>=0.37.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ruff>=0.14.0",
]
